feat/sword Change SpeedSword to DashSword and change dir
