+++ src/main/java/fr/loanspac/magicstuff/MagicStuff.java
@@ -1,7 +1,7 @@
 package fr.loanspac.magicstuff;
 
+import fr.loanspac.magicstuff.dash.DashSword;
 import fr.loanspac.magicstuff.listeners.PlayerListener;
-import fr.loanspac.magicstuff.sword.SpeedSword;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
 
@@ -10,7 +10,7 @@ public final class MagicStuff extends JavaPlugin {
     @Override
     public void onEnable() {
         Bukkit.getPluginManager().registerEvents(new PlayerListener(), this);
-        Bukkit.getPluginManager().registerEvents(new SpeedSword(), this);
+        Bukkit.getPluginManager().registerEvents(new DashSword(), this);
     }
 
     @Override
