+++ src/main/java/fr/loanspac/magicstuff/listeners/PlayerListener.java
@@ -1,7 +1,6 @@
 package fr.loanspac.magicstuff.listeners;
 
-import fr.loanspac.magicstuff.sword.SpeedSword;
-import net.kyori.adventure.text.Component;
+import fr.loanspac.magicstuff.dash.DashSword;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
@@ -10,7 +9,6 @@ public class PlayerListener implements Listener {
 
     @EventHandler
     public void onJoin(PlayerJoinEvent event) {
-        event.getPlayer().sendMessage(Component.text("§3Voici votre épée !"));
-        event.getPlayer().getInventory().addItem(new SpeedSword().buildItem());
+        event.getPlayer().getInventory().addItem(new DashSword().buildItem());
     }
 }
